Tweaking jackknife variance at recommendation of @swager
The row-wise variances can dip below zero (after correction).  That
sets a length scale in the variance distribution that can be used to
adjust the values up to avoid negative values while retaining some
semblance of signal.

The empirical Bayes correction [here](
https://github.com/swager/randomForestCI/blob/master/R/calibrateEB.R)
would be better, but this will do for now.