Added RandomForestClassifier
Refactored the RandomForest model to create a base randomforest and
a base lololearner. Not quite happy with the fact we have to check
whether the concrete instance of the class is a regressor or classifier,
but it isn't too unreasonable because it maximizes the amount of code
reuse between classes. I'll ponder more