Added uber jar profile to pom.xml with shade plugin conf.
The idea is to generate an uber (fat/shaded/..) jar during packaging and attach it to maven reactor goal so that it is uploaded to central.

Though, to keep existing build process clean and as it is, added a new profile that can be activated intentionally (opt-in).

Also, minor version bump due to changes.