Refactor to split impls by type
The MultiBaggedResult was riddled with if/else logic to handle both
regression and classification in the same class.  This was more confusing
than it was deduplicating, because we don't jackknife for classification.

This splits the classification result out explicitly, making it much
simpler and allowing the regression case to have the correct
T = Double generic class.