Update lolopy tests
useJackknife doesn't really do anything at fit time anymore, so
that test was simply removed

Because of that, you cannot get constant predictive uncertainty anymore,
so that test was adjusted to show that modeling the bias only increases
the total predicted uncertainty.

The python interface and tests will need more of an update after more
of these methods are implemented (next PR).