Refactoring the shapley impl to be easier to follow
The major changes here are to split the tracking of the
weights-by-subset-size from the features, and to use a set to
store the information about the features (since their order doesn't matter).

This includes several renamings and changes docstrings.  It also adds a few
illustrative requires statements to document the constraints on values.